Verify sources files can be read and written
We ignore errors when initially pulling form the remote because the
remote files may not exists yet. However, if the local files exists and
aren't writable we will also get errors and in this case instead of
being harmless, ignoring these can result in mass deletions. Now we
pre-verify that the top-level local files are read/writable